Psalm fixes (#4)
* Fix Psalm issues
