+++ lib/Auth/Process/AttributeAddFromLDAP.php
@@ -2,7 +2,7 @@
 
 namespace SimpleSAML\Module\ldap\Auth\Process;
 
-use SimpleSAML\module\ldap\Auth\LDAP;
+use SimpleSAML\Module\ldap\Auth\Ldap;
 
 use Webmozart\Assert\Assert;
 
@@ -157,7 +157,7 @@ class AttributeAddFromLDAP extends BaseFilter
             $arrSearch[] = '%'.$attr.'%';
 
             if (strlen($val[0]) > 0) {
-                $arrReplace[] = LDAP::escape_filter_value($val[0]);
+                $arrReplace[] = Ldap::escape_filter_value($val[0]);
             } else {
                 $arrReplace[] = '';
             }
