+++ lib/Auth/Process/BaseFilter.php
@@ -2,7 +2,7 @@
 
 namespace SimpleSAML\Module\ldap\Auth\Process;
 
-use SimpleSAML\module\ldap\Auth\LDAP;
+use SimpleSAML\Module\ldap\Auth\Ldap;
 
 /**
  * This base LDAP filter class can be extended to enable real
@@ -51,7 +51,7 @@ abstract class BaseFilter extends \SimpleSAML\Auth\ProcessingFilter
      * Instance, object of the ldap connection. Stored here to
      * be access later during processing.
      *
-     * @var \SimpleSAML\Module\ldap\Auth\LDAP|null
+     * @var \SimpleSAML\Module\ldap\Auth\Ldap|null
      */
     private $ldap = null;
 
@@ -257,7 +257,7 @@ abstract class BaseFilter extends \SimpleSAML\Auth\ProcessingFilter
      * rather than setting in the constructor to avoid unnecessarily
      * connecting to LDAP when it might not be needed.
      *
-     * @return \SimpleSAML\Module\ldap\Auth\LDAP
+     * @return \SimpleSAML\Module\ldap\Auth\Ldap
      */
     protected function getLdap()
     {
@@ -290,7 +290,7 @@ abstract class BaseFilter extends \SimpleSAML\Auth\ProcessingFilter
         );
 
         // Connect to the LDAP server to be queried during processing
-        $this->ldap = new LDAP($hostname, $enable_tls, $debug, $timeout, $port, $referrals);
+        $this->ldap = new Ldap($hostname, $enable_tls, $debug, $timeout, $port, $referrals);
         $this->ldap->bind($username, $password);
 
         // All done
